Update usetokensupply.mdx
According to the docs, the `useTokenSupply` hook expects a parameter of type  `<TokenContract>`. Using the address will not work (in a typescript/safe environment at least). I updated it to instantiate the contract instance first by using `useContract` then passing the contract itself in as the expected parameter for `useTokenSupply`.


Signed-off-by: 0xD1x0n  <131799358+devbydixon@users.noreply.github.com>